requirements_accurate.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.5.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Geospatial libraries
geopy>=2.3.0
folium>=0.14.0

# API clients
googlemaps>=4.10.0
mapbox>=0.18.0

# Computer vision for roof detection
opencv-contrib-python>=4.8.0
scikit-learn>=1.3.0

# Optional: For advanced image analysis
# tensorflow>=2.13.0  # Uncomment if using TensorFlow models
# torch>=2.0.0        # Uncomment if using PyTorch models
